Add useQuery
